% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dslcstyle-package.R
\docType{package}
\name{dslcstyle-package}
\alias{dslcstyle}
\alias{dslcstyle-package}
\title{dslcstyle: Themes and Styles for the Data Science Learning Community}
\description{
For now this holds the logo code and some colors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dslc-io/dslcstyle}
  \item \url{https://dslc-io.github.io/dslcstyle/}
  \item Report bugs at \url{https://github.com/dslc-io/dslcstyle/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
